pyproject.toml
[tool.poetry]
name = "simple-blog"
version = "0.1.0"
description = "A simple blog on python"
authors = ["Nickita Rusanov <nipe322@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-dotenv = "^0.5.2"
httpx = "^0.26.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"
env_files = [".test.env"]